+++ src/components/Input/styles.scss
@@ -1,4 +1,5 @@
 $text-color: #fff;
+
 $linear-gradient-yellow: linear-gradient(
   180deg,
   rgba(252, 240, 0, 0) -3.3%,
@@ -14,9 +15,12 @@ $linear-gradient-pink: linear-gradient(
   rgba(246, 43, 253, 0) -3.3%,
   rgba(246, 43, 253, 0.3) 100%
 );
+
 $color-border-botton-yellow: #fcf000;
 $color-border-botton-blue: #00edeb;
 $color-border-botton-pink: #f62bfd;
+
+$box-shadow-blue: rgba(31, 203, 255, 0.53);
 $box-shadow-pink: rgba(246, 43, 253, 0.65);
 
 .textbox {
@@ -24,31 +28,14 @@ $box-shadow-pink: rgba(246, 43, 253, 0.65);
   height: 42px;
   position: relative;
   transform-style: preserve-3d;
-
-  &Box {
-    background-color: transparent;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    // box-shadow: inset 0 0 0 3px #272343
-    position: relative;
-    height: 100%;
-    transform-style: preserve-3d;
-    transition: all 0.2s ease;
-  }
-
-  &Field {
-    display: flex;
-    flex-direction: column;
-    padding: 0px 2px;
-    align-self: stretch;
-    width: 100%;
-  }
+  background-color: transparent;
+  transition: all 0.2s ease;
 
   &Text {
     width: 100%;
     height: 100%;
     border: none;
+    padding: 0 2px;
     background-color: transparent;
     font-size: 1.25rem;
     display: block;
@@ -66,40 +53,6 @@ $box-shadow-pink: rgba(246, 43, 253, 0.65);
     }
   }
 
-  &Face {
-    position: absolute;
-    background: $linear-gradient-blue;
-    left: 0;
-  }
-
-  &Triangles {
-    display: none;
-    // height: 100%;
-    // width: 100%;
-    // bottom: 0;
-    // background: transparent;
-
-    // &:before {
-    //   content: '';
-    //   border-width: 12px 7px 12px 7px;
-    //   border-color: black transparent transparent black;
-    //   position: absolute;
-    //   left: 0px;
-    //   bottom: 0;
-    //   border-style: solid;
-    // }
-
-    // &:after {
-    //   content: '';
-    //   border-width: 12px 7px 12px 7px;
-    //   border-color: black black transparent transparent;
-    //   position: absolute;
-    //   right: 0px;
-    //   bottom: 0;
-    //   border-style: solid;
-    // }
-  }
-
   &BottomGradient {
     height: 35px;
     width: 100%;
@@ -108,14 +61,36 @@ $box-shadow-pink: rgba(246, 43, 253, 0.65);
     transform: perspective(100px) rotateX(50deg);
   }
 
+  &Face {
+    position: absolute;
+    left: 0;
+
+    &Blue {
+      background: $linear-gradient-blue;
+    }
+
+    &Pink {
+      background: $linear-gradient-pink;
+    }
+  }
+
   &BottomLine {
+    position: absolute;
     height: 2px;
     width: 100%;
     bottom: 0;
     transform-origin: 50% 0;
-    border: 1px solid $color-border-botton-blue;
     border-radius: 2px;
     box-sizing: border-box;
-    box-shadow: 0px 0px 6px 2px rgba(31, 203, 255, 0.53);
+
+    &Blue {
+      border: 1px solid $color-border-botton-blue;
+      box-shadow: 0px 0px 6px 2px $box-shadow-blue;
+    }
+
+    &Pink {
+      border: 1px solid $color-border-botton-pink;
+      box-shadow: 0px 0px 6px 2px $box-shadow-pink;
+    }
   }
 }
