+++ src/components/Select/stylesLinear.scss
@@ -1,0 +1,47 @@
+
+$text-color: #fff;
+$linear-gradient-yellow: linear-gradient(
+  180deg,
+  rgba(252, 240, 0, 0) -3.3%,
+  rgba(252, 240, 0, 0.25) 100%
+);
+$linear-gradient-blue: linear-gradient(
+  180deg,
+  rgba(0, 237, 235, 0) -3.3%,
+  rgba(0, 237, 235, 0.35) 100%
+);
+$color-border-botton-yellow: #fcf000;
+$color-border-botton-blue: #00edeb;
+
+.textbox {
+  width: 100%;
+  height: 42px;
+  position: relative;
+  transform-style: preserve-3d;
+
+  &Face {
+    position: absolute;
+    background: $linear-gradient-yellow;
+    left: 0;
+  }
+
+
+  &BottomGradient {
+    height: 35px;
+    width: 100%;
+    bottom: 2px;
+    transform-origin: 50% 100%;
+    transform: perspective(100px) rotateX(50deg);
+  }
+
+  &BottomLine {
+    height: 2px;
+    width: 100%;
+    bottom: 0;
+    transform-origin: 50% 0;
+    border: 1px solid $color-border-botton-yellow;
+    border-radius: 2px;
+    box-sizing: border-box;
+    box-shadow: 0px 0px 6px 2px rgba(31, 203, 255, 0.53);
+  }
+}
