+++ src/containers/governance/actionBarDatail.jsx
@@ -22,6 +22,7 @@ import {
   CheckAddressInfo,
   ButtonImgText,
   Account,
+  Input,
 } from '../../components';
 import { AppContext } from '../../context';
 
@@ -37,7 +38,6 @@ import {
   DEFAULT_GAS_LIMITS,
   VOTE_OPTION,
 } from '../../utils/config';
-import Input from '../teleport/components/input';
 import BtnGrd from '../../components/btnGrd';
 
 const imgKeplr = require('../../image/keplr-icon.svg');
