+++ src/containers/ipfsSettings/index.jsx
@@ -3,7 +3,6 @@ import { connect } from 'react-redux';
 import { ContainerGradientText, MainContainer } from '../portal/components';
 import BtnPasport from '../portal/pasport/btnPasport';
 import Select from '../teleport/components/select';
-import Input from '../teleport/components/input';
 import {
   updateIpfsStateUrl,
   updateIpfsStateType,
@@ -14,6 +13,7 @@ import {
 import InfoIpfsNode from './ipfsComponents/infoIpfsNode';
 import PendingIpfsSettings from './PendingIpfsSettings';
 import ErrorIpfsSettings from './ErrorIpfsSettings';
+import { Input } from '../../components';
 
 const dataOpts = ['external', 'embedded'];
