+++ src/containers/portal/stateComponent/InputNickname.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import Input from '../../teleport/components/input';
 import { ContainerGradient } from '../components';
 import { steps } from '../citizenship/utils';
+import { Input } from '../../../components';
 
 function InputNickname({ valueNickname, step, onChangeNickname }) {
   return (
