+++ src/containers/teleport/components/index.js
@@ -2,6 +2,5 @@ import Select from './select';
 import BalanceToken from './balanceToken';
 import TokenSetter from './tokenSetter';
 import TabList from './tabList';
-import Input from './input';
 
-export { Select, BalanceToken, TokenSetter, TabList, Input };
+export { Select, BalanceToken, TokenSetter, TabList };
