--- src/containers/teleport/components/input/stylesLinear.scss
@@ -1,94 +1,0 @@
-
-$text-color: #fff;
-$linear-gradient-yellow: linear-gradient(
-  180deg,
-  rgba(252, 240, 0, 0) -3.3%,
-  rgba(252, 240, 0, 0.25) 100%
-);
-$linear-gradient-blue: linear-gradient(
-  180deg,
-  rgba(0, 237, 235, 0) -3.3%,
-  rgba(0, 237, 235, 0.35) 100%
-);
-$color-border-botton-yellow: #fcf000;
-$color-border-botton-blue: #00edeb;
-
-.textbox {
-  width: 100%;
-  height: 42px;
-  position: relative;
-  transform-style: preserve-3d;
-
-  &Box {
-    background-color: transparent;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    // box-shadow: inset 0 0 0 3px #272343
-    position: relative;
-    height: 100%;
-    transform-style: preserve-3d;
-    transition: all 0.2s ease;
-  }
-
-  &Field {
-    display: flex;
-    flex-direction: column;
-    padding: 0px 2px;
-    align-self: stretch;
-  }
-
-  &Face {
-    position: absolute;
-    background: $linear-gradient-yellow;
-    left: 0;
-  }
-
-  &Triangles {
-    height: 100%;
-    width: 100%;
-    bottom: 0;
-    background: transparent;
-
-    &::before {
-      content: '';
-      border-width: 12px 7px 12px 7px;
-      border-color: black transparent transparent black;
-      position: absolute;
-      left: 0px;
-      bottom: 0;
-      border-style: solid;
-      z-index: -1;
-    }
-
-    &::after {
-      content: '';
-      border-width: 12px 7px 12px 7px;
-      border-color: black black transparent transparent;
-      position: absolute;
-      right: 0px;
-      bottom: 0;
-      border-style: solid;
-      z-index: -1;
-    }
-  }
-
-  &BottomGradient {
-    height: 35px;
-    width: 100%;
-    bottom: 2px;
-    transform-origin: 50% 100%;
-    transform: perspective(100px) rotateX(50deg);
-  }
-
-  &BottomLine {
-    height: 2px;
-    width: 100%;
-    bottom: 0;
-    transform-origin: 50% 0;
-    border: 1px solid $color-border-botton-yellow;
-    border-radius: 2px;
-    box-sizing: border-box;
-    box-shadow: 0px 0px 6px 2px rgba(31, 203, 255, 0.53);
-  }
-}
